Add simple 1 level priority inheritance support.  Improve some queue APIs -- make them more consistent.

+++ mos/defs.h
@@ -56,6 +56,6 @@ typedef int32_t     s32;
 typedef uint64_t    u64;
 typedef int64_t     s64;
 
-typedef u16 MosThreadPriority;
+typedef u8 MosThreadPriority;
 
 #endif
